requirements.txt
absl-py==2.1.0
accelerate==1.1.1
addict==2.4.0
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
anyio==4.6.2.post1
astunparse==1.6.3
attrs==23.2.0
basicsr==1.4.2
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cog==0.13.3
coloredlogs==15.0.1
contourpy==1.3.1
controlnet-aux==0.0.9
customtkinter==5.2.2
cycler==0.12.1
Cython==3.0.11
darkdetect==0.8.0
diffusers==0.31.0
easydict==1.13
einops==0.8.0
eval_type_backport==0.2.0
facexlib==0.3.0
fastapi==0.110.3
filelock==3.16.1
filterpy==1.4.5
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.0
fsspec==2024.10.0
future==1.0.0
gast==0.6.0
gdown==5.2.0
gfpgan==1.3.8
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.3
humanfriendly==10.0
idna==3.10
imageio==2.36.1
importlib_metadata==8.5.0
insightface==0.7.3
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.7.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lmdb==1.5.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mediapipe==0.10.18
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.2
onnx==1.17.0
onnxruntime==1.20.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opennsfw2==0.14.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
prettytable==3.12.0
protobuf==4.25.5
psutil==6.1.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
PyMatting==1.1.13
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
rembg==2.0.60
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
setuptools==75.6.0
simsimd==6.2.1
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
starlette==0.37.2
stringzilla==3.11.0
structlog==24.4.0
sympy==1.13.1
tb-nightly==2.19.0a20241121
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.6.7
tokenizers==0.20.3
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
yapf==0.43.0
zipp==3.21.0
